+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- 以下権限を追加
+    - `events.*`
+    - `events.read`
+    - `orders.read`
+    - `orders.findByConfirmationNumber`
+
 ### Deprecated
 
 ### Removed